Moves Proxy over to PB endpoint
