+++ sokannonser/__init__.py
@@ -5,8 +5,7 @@ from sokannonser import appconf
 from sokannonser.rest import api
 # Import all Resources that are to be made visible for the app
 from sokannonser.rest.endpoint.auranest import AuranestSearch
-from sokannonser.rest.endpoint.platsannonser import PBSearch
-# from sokannonser.rest.endpoint.openapi import Proxy
+from sokannonser.rest.endpoint.platsannonser import PBSearch, Proxy
 # from sokannonser.rest.endpoint.openapi import OpenSearch
 from sokannonser.rest.endpoint.valuestore import Valuestore
