+++ sokannonser/rest/__init__.py
@@ -6,7 +6,7 @@ api = Api(version='1.0', title='Search job ads',
           default='sokannonser',
           default_label="An API for searching and retrieving job ads.")
 
-ns_platsannons = Namespace('AF-job ads',
+ns_platsannons = Namespace('Open AF-job ads',
                            description='Search and retrieve Arbetsförmedlingens (AF) '
                            'job ads. Used for online operations.')
 ns_auranest = Namespace('All job ads',
