+++ sokannonser/rest/endpoint/openapi.py
@@ -1,9 +1,7 @@
 import logging
-from flask_restplus import Resource, abort
-from requests import get, exceptions
+from flask_restplus import Resource
 from jobtech.common.rest.decorators import check_api_key
-from sokannonser import settings
-from sokannonser.rest import ns_platsannons, ns_open
+from sokannonser.rest import ns_open
 from sokannonser.rest.model.platsannons_results import simple_lista
 from sokannonser.rest.model.queries import pb_query
 from sokannonser.rest.model.queries import swagger_doc_params, swagger_filter_doc_params
@@ -36,28 +34,3 @@ class OpenSearch(Resource):
     @ns_open.marshal_with(simple_lista)
     def marshal_results(self, result):
         return result
-
-
-@ns_open.route('/ad/<id>')
-class Proxy(Resource):
-
-    @ns_platsannons.doc(
-        responses={
-            200: 'OK',
-            401: 'Invalid API-key',
-            404: 'Job ad not found exception',
-            500: 'Technical exception'
-        }
-    )
-    def get(self, id):
-        url = "%s%s" % (settings.AD_PROXY_URL, id)
-        headers = {'Accept-Language': 'sv', 'Accept': 'application/json'}
-        try:
-            response = get(url, headers=headers, timeout=10)
-            if response.status_code == 200:
-                return response.json()
-            else:
-                abort(response.status_code)
-        except exceptions.RequestException as e:
-            log.error('Failed to connect', e)
-            abort(500)
